Start of WebCrawler.
Here is a start to it.  All it does now is takes a URL, goes and looks
at HTML.  Finds a href tag, and prints it.  We need to integrate the UI,
and then write to a text file to store the found URLs.  Then we can
check stored URL's against new ones to make sure we don't loop through
the same one over and over.